[Feat] 타이핑 완료 또는 시간 종료 시 결과 화면에 타이핑 결과 반영

+++ DailyTyping/DailyTyping/Sources/Model/PilsaPerformance.swift
@@ -9,6 +9,6 @@ import Foundation
 
 struct PilsaPerformance {
     let wpm: Int
-    let acc: Double
+    let acc: Int
     let date: Date
 }
